requirements.txt
click==5.0
github3.py==0.9.6
PyYAML==3.12
requests==2.18.4
